By default, have debug logs not show in tests.
Previously, debugging logs were shown in tests and had to be shut
off.  This lead to very verbose output that hardly ever helped much.
Now the behavior is the opposite.  Debug logs are not shown by default
but can be turned on with a flag (assuming compiled in debug mode).