Fix MSVC debugger warning for not initalizing function argument.
The latest MSVC debugger apparently checks to make sure that any
variable used as an argument to a function has been set.  There was
an instance in the reduce strategy that violated it.  I don't think
it could ever cause a problem (the arguments are ignored in this
particular state), but it's still bad programming style.  Thanks
to John Biddiscombe for bringing this up.