Set CMake policy 0017 to old behavior.
This behavior makes includes in modules shipped with CMake still look
in the module directories defined by the project.  The change does not
effect building IceT on its own.  (At least, there are currently no
modules included that are also included with CMake.)  However, this is
currently necessary to prevent warnings in ParaView.