Make RandomTransform test less susceptible to numerical error.
The way the translation transform was picked, I was getting z shifts
that were so close that after OpenGL transforms you ended up with the
same depth value, which of course threw off the compositing.  This
change limits the precision used for the Z shift which seems to avoid
really close values but does not seem to encourage the same value
(although that is still a possibility).