Work around when MPI_IN_PLACE not available.
The MPI_IN_PLACE feature, which allows you to do collective operations
such as gather with the local data already in the destination array, is
a feature that is not available in MPI 1.  To continue to support MPI
version 1, implement an alternative copying version that is only used
for that version.