Remove dependence on GLU
GLU seems to be deprecated in OpenGL 3 and now the Apple XCode compile
is warning about using it. The only thing we were using it for was to
draw a sphere, so replace that with our own sphere drawing code (using
vertex arrays that I think are still going to be supported).

An alternative to rolling our own would be to use a canned version that
comes with GLUT. However, GLUT itself is also going through a
transitional period, so until we can be sure it is going to survive and
be available everywhere, this is a reasonable solution.