Support GLFW library for tests.
Previously we were using GLUT to create windows and OpenGL contexts.
However, the base GLUT is no longer maintained and APPLE as deprecated
its use. There is another project, FreeGLUT, that is poised to replace
the original GLUT library, but it does not support native Mac windows
yet.

There is another library named GLFW that is also a cross-platform
library for creating OpenGL windows, and this one supports the native
Mac windows. For now we are supporting both (with the default being
GLUT, which is still more common).