Add an image format for pointers to color/depth buffers
The IceTImage object expects a single allocated buffer in which both the
color and depth is encoded. This is convienient for passing images in a
single message. However, when interfacing with client code, it is
typical to have them in separate buffers. Thus, we have a secondary
dense image format that has pointers to these two buffers. It should be
able to do everything a normal image can do except be packaged for a
send or be written to.