Adjust size of viewport when rendering tiles
In the RandomTransform test when rendering pre-rendered images, we were
previously always rendering to the entire window size. However, when
trying tiled displays, which break up the display, the actual global
display can be a little smaller than that due to roundoff in the
division. This change sets the viewport to be exactly the size of the
global viewport.