Collect timing in SimpleTiming before printing
Previously, the root process would print the log at the end of
SimpleTimingDoRender. The problem is that in the scaling studies the
world communicator is partitioned so that there are multiple root
processes running at once. I've been having trouble with log entries
stomping over each other.

This change holds all the timings in a buffer rather than print them.
After scaling studies are complete, all log entries are collected to the
world root and printed out there.