Remove reporting of no error in tests
At the end of each test each process reported whether it got an IceT or
OpenGL error including if it is no error. That can be helpful in part to
know that the process successfully completed, but for tests on very
large processes it generated excessively large log files. Instead
suppress the report if there is no error.