Make ranks argument of icetCommSubset constant
The subset command of the communicator takes an array of ranks, but that
array should not be changed. Declare that array const in all versions of
the function. (This requires a const cast when calling the MPI version
since that function does not take a const array even though it should.)