Separate out OpenGL and non-OpenGL tests
Previously, if OpenGL was enabled all tests would create a render window
even if it was never used because they were all crammed into the same
executable that initalized with the OpenGL context. This is an artifact
from early versions of IceT that always required OpenGL.

Creating an OpenGL context shouldn't change the results of the test, but
it was annoying to have a bunch of windows created for every test. This
change separates out the OpenGl tests from those that don't use it so
that most of the tests now just run in the background.