Add new radix-kr single image composite algorithm
The radix-kr is essentially the radix-k algorithm with some ideas from
2-3 swap added in. Rather than forcing all the k's in radix-k to be
factors, radix-kr allows the k to have a remainder when splitting the
process. The remaining processes participate in that round by splitting
their image and sending them off, but they receive nothing that round
and then go idle. Although this adds some imbalance, it does not add
much. It also makes keeping track of partitions easier than 2-3 swap.