Add dense images option to SimpleTiming.
IceT insists on using sparse images most of the time because they are
almost always better (at least at the beginning part of the
compositing). Add an option to SimpleTiming to draw images that have no
background pixels so that the "sparse" images are equivalent to dense
images. This is done by making the depth of background pixels slightly
above and the opacity never 0.