Update the diagnostics macros
The new version now uses variadic template macros to allow you to do
printf-style messages. This is how I wanted it to begin with, but as I
recall variadic templates were not supported on all compilers at the
time I started IceT. At this point, I think they are supported on pretty
much any compiler in the last 10 years.