Make default ICET_MAX_IMAGE_SPLIT higher
Move the default ICET_MAX_IMAGE_SPLIT from 512 to a much higher 500000.
The initial value was based on some timings on Intrepid (a BlueGene).
However, I have not seem the same benefit on other architectures, so in
the spirit of do no harm I'm raising this back up. Perhaps with some
more data I will move it back down again.