Add a render layer for OpenGL 3
Added a new library for interfacing IceT with OpenGL version 3. A lot
has changed with OpenGL since IceT was first written. This device layer
relies less on a fixed function pipeline and instead uses the OpenGL
device to transfer data back and forth more than anything.

The current render creates OpenGL textures and passes their identifiers
to the callback. The callback is expected to fill those textures with
the rendering.

Other types of callback could be created in the future.