Pass framebuffer instead of textures to OGL3 callback
Rather than sending textures to the OpenGL 3 callback and requiring the
callback to rebuild the framebuffer every time, create the framebuffer
internally inside of IceT and pass the framebuffer id to the callback.

I am also binding the framebuffer before the call. I don't know if it is
necessary to both set the framebuffer and pass the framebuffer id.