Allow render layer to manage retrieval and compression
The previous version of IceT always had the render layer return
an uncompressed image buffer, which was used to later compress
or otherwise manage. Eventually, we would like to have images
compressed on the graphics card, and therefore it is up to the
render layer to do this compression. Thus, change the image
retreival to allow the render layer to do this.