Create larger OpenGL 3 render buffers
When rendering with multiple tiles, IceT's floating viewport only works
if the geometry fits within the rendering buffer. This is much more
likely if the rendering buffer is bigger.

The OpenGL 3 rendering layer uses textures for an offscreen rendering
buffer. We can make this buffer as large as we want (given constraints
of the rendering system). As such, make the rendering buffer the same
size as the whole display (or as large as possible if that is not
possible).